bug fix: output <meta name="description"> tags using largo_opengraph logic, because Yoast SEO doesn't.
This continues the work in https://github.com/INN/umbrella-mstoday/pull/13,
addressing specifically the fact that the Yoast SEO plugin outputs
twitter:description and og:description tags, but not plain description tags.

So here we copy a bunch of logic from largo_opengraph, and run it
if Yoast is active.